+++ docs/isaacsim/snippets/replicator_tutorials/tutorial_replicator_augmentation/annotator_augmentation_script_editor.py
@@ -21,7 +21,7 @@ import carb.settings
 import numpy as np
 import omni.replicator.core as rep
 import warp as wp
-from isaacsim.core.utils.stage import open_stage
+from isaacsim.core.experimental.utils.stage import open_stage
 from isaacsim.storage.native import get_assets_root_path_async
 from omni.replicator.core.functional import write_image