+++ source/extensions/isaacsim.core.experimental.utils/python/mock/bounds.py
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: Copyright (c) 2021-2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,8 +13,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Base sample framework for creating interactive Isaac Sim examples with UI templates."""
+"""Bounds module."""
 
-
-from .base_sample import BaseSample
-from .base_sample_extension import BaseSampleUITemplate
+from .impl.bounds import *